+++ tests/value_stability.rs
@@ -502,11 +502,13 @@ fn weibull_stability() {
 fn dirichlet_stability() {
     let mut rng = get_rng(223);
     assert_eq!(
-        rng.sample(Dirichlet::new([1.0, 2.0, 3.0]).unwrap()),
+        multi::Dirichlet::new([1.0, 2.0, 3.0])
+            .unwrap()
+            .sample(&mut rng),
         [0.12941567177708177, 0.4702121891675036, 0.4003721390554146]
     );
     assert_eq!(
-        rng.sample(Dirichlet::new([8.0; 5]).unwrap()),
+        multi::Dirichlet::new([8.0; 5]).unwrap().sample(&mut rng),
         [
             0.17684200044809556,
             0.29915953935953055,
@@ -517,7 +519,9 @@ fn dirichlet_stability() {
     );
     // Test stability for the case where all alphas are less than 0.1.
     assert_eq!(
-        rng.sample(Dirichlet::new([0.05, 0.025, 0.075, 0.05]).unwrap()),
+        multi::Dirichlet::new([0.05, 0.025, 0.075, 0.05])
+            .unwrap()
+            .sample(&mut rng),
         [
             0.00027580456855692104,
             2.296135759821706e-20,